Refactor to break down large source files
